requirements.txt
streamlit
crewai
langchain
langchain-community
langchain-openai
openai
google-search-results
langfuse
python-dotenv
pydantic
requests
anthropic